That's me, a truly outsider in this place
I tried to learn Python on pandemics times, but little advance on basic ideas was then possible. Nowadays I find it difficult to resume this task, but some ideas deserving to be translated to a commands line might reach its blank by joining forces. Stepwise presentation of my parole-fashioned procedures in order to achieve codes to be tested calls for the patience of expert developers. Final outcome should be the very first fully deterministic solution of TSP. Believe it, or not.